requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
python-dotenv==1.0.0
pytesseract==0.3.10
pdf2docx==0.5.6
python-docx==1.1.0
requests==2.31.0
alembic==1.12.1
python-multipart==0.0.6
Pillow==10.1.0
websockets==12.0
starlette==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
typing-extensions==4.8.0